[PPAOPS-3701] Add command line options --sgfs and --sgfr
These options allow to skip calls to vkWaitForFences and vkGetFenceStatus.
The '--sgfs <status>' option (or --skip-get-fence-status) sets the condition
on which the calls are skipped (never, unsuccessful, always).
The '--sgfr <ranges>' option (or --skip-get-fence-ranges) sets the frame ranges
where the skip condition is applied (if not specified, applies to all frames).

Change-Id: I235b46530b7ec6e98d5a0432b68990b37cc851f4