Revert measurements file commits
Revert commits:
- 9b3a1a2a30f7e24076a2c8177bebbdbb8ea9ddf3
- 1d37484f491287820b33c2e52816463cfb730d5f
- 7c296cfbc897fce5903838f789644f48a103bf2e

LunarG guys wants some modification to this feature and might
request more. So the idea is that this commit creates a "clean
point" from which I will reimplement the feature in a way everyone
is pleased in future commits. If I need some modifications, I can
then revert these future commits to get back to this "clean point".

Git Hell.

Change-Id: Ic06dbcf75509ac24d972fa34ce6dc6d5b6544ae6